Made the Java version use a char, and the C++ version use a uint16_t for optimization

+++ cpp/Quadrant/src/quadrant.hpp
@@ -58,7 +58,7 @@ private:
 
 private:
     // Data
-    int kbs;
+    uint16_t kbs;
 
     // Constants
     static inline constexpr int KLINGON_MAX = 3;
